% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gene_catalog.R
\name{build_gene_catalog}
\alias{build_gene_catalog}
\title{Build full human gene catalog}
\usage{
build_gene_catalog(outdir = ".")
}
\arguments{
\item{outdir}{Output directory}
}
\value{
A data.frame with the gene catalog
}
\description{
Download, process, and integrate NCBI gene resources
into a unified human gene catalog.
}
