% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_gene_catalog_data}
\alias{download_gene_catalog_data}
\title{Download gene catalog source files}
\usage{
download_gene_catalog_data(path = tools::R_user_dir("MAHGIx", which = "cache"))
}
\description{
Download gene catalog source files
}
