% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ncbi.R
\name{download_ncbi_data}
\alias{download_ncbi_data}
\title{Download NCBI human gene resources}
\usage{
download_ncbi_data()
}
\value{
A named list of data.tables
}
\description{
Download NCBI human gene resources
}
