% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_human.R
\name{filter_human_9606}
\alias{filter_human_9606}
\title{Filter human genes (tax_id = 9606)}
\usage{
filter_human_9606(gene_info, gene_neighbors, gene2accession)
}
\arguments{
\item{gene_info}{NCBI gene_info table}

\item{gene_neighbors}{gene_neighbors table}

\item{gene2accession}{gene2accession table}
}
\value{
A list of filtered tables
}
\description{
Filter human genes (tax_id = 9606)
}
