% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_inspect.R
\name{inspect_gene_catalog}
\alias{inspect_gene_catalog}
\title{Provides an extended overview of the gene catalog structure and annotation coverage.}
\usage{
inspect_gene_catalog(catalog)
}
\arguments{
\item{catalog}{Gene catalog generated with build_gene_catalog()}
}
\description{
Provides an extended overview of the gene catalog structure and annotation coverage.
}
