% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rutgers_map.R
\name{map_rutgers_positions}
\alias{map_rutgers_positions}
\title{Map physical to genetic positions (Rutgers Map v3)}
\usage{
map_rutgers_positions(df, rutgers_dir)
}
\arguments{
\item{df}{Gene catalog dataframe}

\item{rutgers_dir}{Directory with Rutgers map files}
}
\value{
Data frame with genetic positions
}
\description{
Map physical to genetic positions (Rutgers Map v3)
}
