% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_inspect.R
\name{summary_gene_catalog}
\alias{summary_gene_catalog}
\title{Provides a quick overview of the gene catalog contents.}
\usage{
summary_gene_catalog(catalog)
}
\arguments{
\item{catalog}{Gene catalog generated with build_gene_catalog()}
}
\description{
Provides a quick overview of the gene catalog contents.
}
