% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_catalog.R
\name{update_gene_catalog}
\alias{update_gene_catalog}
\title{Update gene catalog}
\usage{
update_gene_catalog(path = tools::R_user_dir("MAHGIx", which = "cache"))
}
\description{
Forces rebuild of the catalog by deleting the cached version.
}
