% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_catalog.R
\name{build_gene_catalog}
\alias{build_gene_catalog}
\title{Build integrated human gene catalog}
\usage{
build_gene_catalog(path = tools::R_user_dir("MAHGIx", which = "cache"))
}
\arguments{
\item{path}{Cache directory}
}
\description{
Downloads required datasets if missing, builds the catalog,
and caches the result for fast reuse.
}
